requirements.txt
# EV Charging Simulation Requirements

# OCPP Protocol
ocpp>=0.29.0
websockets>=12.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Dashboard
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
python-dateutil>=2.8.0
scipy>=1.10.0
joblib>=1.3.0
